requirements.txt
streamlit
tensorflow-cpu
librosa
noisereduce
pillow
matplotlib